# openapi.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/profile/update":  {
                                              "put":  {
                                                          "tags":  [
                                                                       "profile-controller"
                                                                   ],
                                                          "operationId":  "updateProfile",
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  "@{schema=}"
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  "@{*/*=}"
                                                                                    }
                                                                        }
                                                      }
                                          },
                  "/api/admin/unblock/{userId}":  {
                                                      "put":  {
                                                                  "tags":  [
                                                                               "admin-controller"
                                                                           ],
                                                                  "operationId":  "unblockUser",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "userId",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  "@{type=integer; format=int64}"
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "OK",
                                                                                                "content":  "@{*/*=}"
                                                                                            }
                                                                                }
                                                              }
                                                  },
                  "/api/admin/block/{userId}":  {
                                                    "put":  {
                                                                "tags":  [
                                                                             "admin-controller"
                                                                         ],
                                                                "operationId":  "blockUser",
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "userId",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  "@{type=integer; format=int64}"
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  "@{*/*=}"
                                                                                          }
                                                                              }
                                                            }
                                                },
                  "/api/admin/approve/{userId}":  {
                                                      "put":  {
                                                                  "tags":  [
                                                                               "admin-controller"
                                                                           ],
                                                                  "operationId":  "approveUser",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "userId",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  "@{type=integer; format=int64}"
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "OK",
                                                                                                "content":  "@{*/*=}"
                                                                                            }
                                                                                }
                                                              }
                                                  },
...
                  "/api/profile/complete":  {
                                                "post":  {
                                                             "tags":  [
                                                                          "profile-controller"
                                                                      ],
                                                             "operationId":  "completeProfile",
                                                             "requestBody":  {
                                                                                 "content":  {
                                                                                                 "application/json":  "@{schema=}"
                                                                                             },
                                                                                 "required":  true
                                                                             },
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  "@{*/*=}"
                                                                                       }
                                                                           }
                                                         }
                                            },
...
                  "/api/admin/user-profile/{userId}":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "admin-controller"
                                                                                ],
                                                                       "operationId":  "getUserProfile",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "userId",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  "@{type=integer; format=int64}"
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK",
                                                                                                     "content":  "@{*/*=}"
                                                                                                 }
                                                                                     }
                                                                   }
                                                       },
...
                  "/api/admin/delete/{userId}":  {
                                                     "delete":  {
                                                                    "tags":  [
                                                                                 "admin-controller"
                                                                             ],
                                                                    "operationId":  "deleteUser",
                                                                    "parameters":  [
                                                                                       {
                                                                                           "name":  "userId",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  "@{type=integer; format=int64}"
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "OK",
                                                                                                  "content":  "@{*/*=}"
                                                                                              }
                                                                                  }
                                                                }
                                                 }
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}